refactor: rename social login component to follow naming conventions
